// frontend/src/api/generated/v1/zod/contracts/contracts.ts
export const contractsUpdateBodyNameOneMax = 255;

export const contractsUpdateBodyTotalAmountOneMin = 0;

export const contractsUpdateBodyTotalAmountTwoRegExp = new RegExp('^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$');

export const contractsUpdateBodyDescriptionDefault = ``;

export const ContractsUpdateBody = zod.object({
  "contract_type": zod.union([zod.string(),zod.null()]).optional(),
  "service_tier": zod.union([zod.string(),zod.null()]).optional(),
  "supplier": zod.union([zod.string(),zod.null()]).optional(),
  "client": zod.union([zod.string(),zod.null()]).optional(),
  "name": zod.union([zod.string().min(1).max(contractsUpdateBodyNameOneMax),zod.null()]).optional(),
  "total_amount": zod.union([zod.number().min(contractsUpdateBodyTotalAmountOneMin),zod.string().regex(contractsUpdateBodyTotalAmountTwoRegExp),zod.null()]).optional(),
  "installments_count": zod.union([zod.int().min(1),zod.null()]).optional(),
  "status": zod.union([zod.string(),zod.null()]).optional(),
  "description": zod.string().default(contractsUpdateBodyDescriptionDefault),
  "parent": zod.union([zod.string(),zod.null()]).optional(),
  "pdf_file_key": zod.union([zod.string(),zod.null()]).optional(),
  "expiration_date": zod.union([zod.iso.date(),zod.null()]).optional(),
  "alert_days_before": zod.union([zod.int(),zod.null()]).optional(),
  "signed_date": zod.union([zod.iso.date(),zod.null()]).optional()
}).describe('Schema de entrada para atualização parcial de contrato.')